Move tests to module instead of lib
